requirements.txt
garminconnect==0.2.38
mcp>=1.8.1
python-dotenv>=1.0.0
